// src/DoStuff.Client/DoStuff-Client/src/api/types.gen.ts
export type TimeSettings = {
    id: number;
    key: string;
    userKey: string;
    refreshInterval: number;
    timeZones: Array<TimeZoneSetting>;
};

export type TimeZoneDisplayTime = {
    id: string;
    name?: string | null;
    value?: string | null;
};

export type TimeZoneInfo = {
    readonly id: string;
    readonly hasIanaId: boolean;
    readonly displayName: string;
    readonly standardName: string;
    readonly daylightName: string;
    readonly baseUtcOffset: string;
    readonly supportsDaylightSavingTime: boolean;
};

export type TimeZoneSetting = {
    id: string;
    name?: string | null;
};